Moved most of the transformations.
Only a few things left in AtomGroup. The remaining things in Atom,
Residue, and Segment are either bond stuff or the convenience names
(Segment.<resname>).
