+++ package/MDAnalysis/core/AtomGroup.py
@@ -581,156 +581,6 @@ class AtomGroup(object):
         self.universe.dihedrals += tgroup
         self._clear_caches('dihedrals')
 
-    def rotate(self, R):
-        r"""Apply a rotation matrix *R* to the selection's coordinates.
-
-        AtomGroup.rotate(R)
-
-        :math:`\mathsf{R}` is a 3x3 orthogonal matrix that transforms a vector
-        :math:`\mathbf{x} \rightarrow \mathbf{x}'`:
-
-        .. math::
-
-            \mathbf{x}' = \mathsf{R}\mathbf{x}
-        """
-        R = np.matrix(R, copy=False, dtype=np.float32)
-        # changes the coordinates (in place)
-        x = self.universe.trajectory.ts.positions
-        idx = self.indices
-        x[idx] = x[idx] * R.T  # R.T acts to the left & is broadcasted N times.
-        return R
-
-    def rotateby(self, angle, axis, point=None):
-        r"""Apply a rotation to the selection's coordinates.
-
-        AtomGroup.rotateby(angle,axis[,point])
-
-        The transformation from current coordinates :math:`\mathbf{x}`
-        to new coordinates :math:`\mathbf{x}'` is
-
-        .. math::
-
-          \mathbf{x}' = \mathsf{R}\,(\mathbf{x}-\mathbf{p}) + \mathbf{p}
-
-        where :math:`\mathsf{R}` is the rotation by *angle* around the
-        *axis* going through *point* :math:`\mathbf{p}`.
-
-        :Arguments:
-          *angle*
-             rotation angle in degrees
-          *axis*
-             rotation axis vector, a 3-tuple, list, or array, or a 2-tuple of
-             two MDAnalysis objects from which the axis is calculated as the
-             vector from the first to the second center of geometry.
-          *point*
-             point on the rotation axis; by default (``None``) the center of
-             geometry of the selection is chosen, or, if *axis* is a tuple of
-             selections, it defaults to the first point of the axis. *point*
-             can be a 3-tuple, list, or array or a MDAnalysis object (in which
-             case its :meth:`centroid` is used).
-
-        :Returns: The 4x4 matrix which consists of the rotation matrix ``M[:3,:3]``
-                  and the translation vector ``M[:3,3]``.
-        """
-        alpha = np.radians(angle)
-        try:
-            sel1, sel2 = axis
-            x1, x2 = sel1.centroid(), sel2.centroid()
-            v = x2 - x1
-            n = v / np.linalg.norm(v)
-            if point is None:
-                point = x1
-        except (ValueError, AttributeError):
-            n = np.asarray(axis)
-        if point is None:
-            p = self.centroid()
-        else:
-            try:
-                p = point.centroid()
-            except AttributeError:
-                p = np.asarray(point)
-        M = transformations.rotation_matrix(alpha, n, point=p)
-        self.transform(M)
-        return M
-
-    def align_principal_axis(self, axis, vector):
-        """Align principal axis with index *axis* with *vector*.
-
-        :Arguments:
-          *axis*
-            Index of the principal axis (0, 1, or 2), as produced by
-            :meth:`~MDAnalysis.core.AtomGroup.AtomGroup.principal_axes`.
-          *vector*
-            A 3D vector such as the z-axis (``[0,0,1]``); can be
-            anything that looks like a list with three entries.
-
-        To align the long axis of a channel (the first principal axis,
-        i.e. *axis* = 0) with the z-axis::
-
-          u.atoms.align_principal_axis(0, [0,0,1])
-          u.atoms.write("aligned.pdb")
-        """
-        p = self.principal_axes()[axis]
-        angle = np.degrees(mdamath.angle(p, vector))
-        ax = transformations.rotaxis(p, vector)
-        #print "principal[%d] = %r" % (axis, p)
-        #print "axis = %r, angle = %f deg" % (ax, angle)
-        return self.rotateby(angle, ax)
-
-    def pack_into_box(self, box=None, inplace=True):
-        r"""Shift all atoms in this group to be within the primary unit cell.
-
-        AtomGroup.packintobox([box, [inplace=True]])
-
-        :Keywords:
-          *box*
-            Unit cell to move atoms inside of.
-          *inplace*
-            ``True``: Change coordinates in place and return
-            ``False``: Only return the coordinates
-
-        All atoms will be moved so that they lie between 0 and
-        boxlength :math:`L_i` in all dimensions, i.e. the lower left
-        corner of the simulation box is taken to be at (0,0,0):
-
-        .. math::
-
-           x_i' = x_i - \left\lfloor\frac{x_i}{L_i}\right\rfloor
-
-        The default is to take unit cell information from the
-        underlying :class:`~MDAnalysis.coordinates.base.Timestep`
-        instance. The optional argument *box* can be used to provide
-        alternative unit cell information (in the MDAnalysis standard
-        format ``[Lx, Ly, Lz, alpha, beta, gamma]``).
-
-        Works with either orthogonal or triclinic box types.
-
-        By default the coordinates are changed in place and returned
-
-        .. versionadded:: 0.8
-
-        """
-        if box is None:  #Try and auto detect box dimensions
-            box = self.dimensions  # Can accept any box
-
-        if box.shape == (3, 3):
-            # for a vector representation, diagonal cannot be zero
-            if (box.diagonal() == 0.0).any():
-                raise ValueError("One or more box dimensions is zero."
-                                 "  You can specify a boxsize with 'box ='")
-        else:
-            if (box == 0).any():  #Check that a box dimension isn't zero
-                raise ValueError("One or more box dimensions is zero."
-                                 "  You can specify a boxsize with 'box='")
-
-        coords = self.universe.coord.positions[self.indices]
-        if not inplace:
-            return distances.apply_PBC(coords, box)
-
-        self.universe.coord.positions[self.indices] = distances.apply_PBC(coords, box)
-
-        return self.universe.coord.positions[self.indices]
-
     def wrap(self, compound="atoms", center="com", box=None):
         """Shift the contents of this AtomGroup back into the unit cell.
